Refactor: weight 타입을 float에서 double로 변경 (#22)

+++ src/main/java/com/petplate/petplate/petdailymeal/domain/entity/RawDailyMeal.java
@@ -34,10 +34,10 @@ public class RawDailyMeal {
     private Raw raw;
 
     @Column(nullable = false)
-    private float weight;
+    private double weight;
 
     @Builder
-    public RawDailyMeal(DailyMeal dailyMeal, Raw raw, float weight) {
+    public RawDailyMeal(DailyMeal dailyMeal, Raw raw, double weight) {
         this.dailyMeal = dailyMeal;
         this.raw = raw;
         this.weight = weight;
